synced session info to ui implemented pause resume charging buttons

+++ lib/data/models/session_info.dart
@@ -4,10 +4,10 @@ part 'session_info.g.dart';
 
 @JsonSerializable()
 class SessionInfo {
-  final int charged_energy_wh;
-  final int charging_duration_s;
+  final double charged_energy_wh;
+  final double charging_duration_s;
   final DateTime datetime;
-  final int latest_total_w;
+  final double latest_total_w;
   final String state;
 
   SessionInfo(this.charged_energy_wh, this.charging_duration_s, this.datetime,
